Handle missing dependency when rendering recipe.
The api.render function will raise a DependencyNeedsBuildingError
exception even when "skip: True" is defined for a platform.
This causes the build process to error out.  Catch the exception
and return an empty list so bioconda-utils will proceed.  The
build will then exit nicely with the usual "Nothing to be done"
message as before.